src/CognitiveSearch.Skills/Python/Language/local.requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


azure-functions
azure-functions-durable
numpy
datetime
azure-ai-textanalytics
azure-core
azure-storage-blob>=12.0.0
azure-cosmosdb-table
pandas

#spacy>=3.1.0,<3.2.0
#https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.1.0/en_core_web_md-3.1.0.tar.gz#egg=en_core_web_md
#https://github.com/explosion/spacy-models/releases/download/de_core_news_md-3.1.0/de_core_news_md-3.1.0.tar.gz#egg=de_core_news_md